Add host-min and host-max functions
- We have to include a couple of special cases for /31 and /32
networks. Technically speaking, both of those networks have 0 hosts,
so there is no concept of first or last host in those. We have to
settle to a couple of default values that look consistent with the old
functions, and with a couple of IPv4 calculators I found on the
Internet.